Pull out sql-related code into separate module
